requirements.txt
tinkoff-investments
pandas
numpy
sqlalchemy
python-dotenv
matplotlib